+++ rs/sns/governance/src/governance.rs
@@ -1,12 +1,9 @@
 use crate::extensions::{
-    validate_extension_wasm, ExtensionKind, ExtensionOperationSpec, ExtensionSpec,
-    ValidatedExecuteExtensionOperation, ValidatedRegisterExtension,
+    validate_execute_extension_operation, ExtensionKind, ValidatedExecuteExtensionOperation,
+    ValidatedRegisterExtension,
 };
 use crate::icrc_ledger_helper::ICRCLedgerHelper;
-use crate::pb::sns_root_types::{
-    register_extension_response, CanisterCallError, ListSnsCanistersRequest,
-    ListSnsCanistersResponse,
-};
+use crate::pb::sns_root_types::{register_extension_response, CanisterCallError};
 use crate::pb::v1::governance::GovernanceCachedMetrics;
 use crate::pb::v1::{
     valuation, ExecuteExtensionOperation, Metrics, RegisterExtension, TreasuryMetrics,
@@ -2609,76 +2606,6 @@ impl Governance {
         }
     }
 
-    async fn list_extensions(&self) -> Result<Vec<PrincipalId>, GovernanceError> {
-        let list_extensions_arg = Encode!(&ListSnsCanistersRequest {}).unwrap();
-
-        let ListSnsCanistersResponse { extensions, .. } = self
-            .env
-            .call_canister(
-                self.proto.root_canister_id_or_panic(),
-                "list_extensions",
-                list_extensions_arg,
-            )
-            .await
-            .map_err(|err| {
-                GovernanceError::new_with_message(
-                    ErrorType::External,
-                    format!(
-                        "Canister method call Root.list_extensions failed: {:?}",
-                        err
-                    ),
-                )
-            })
-            .and_then(|blob| {
-                Decode!(&blob, ListSnsCanistersResponse).map_err(|err| {
-                    GovernanceError::new_with_message(
-                        ErrorType::External,
-                        format!("Error decoding Root.list_extensions response: {:?}", err),
-                    )
-                })
-            })?;
-
-        let extensions = extensions
-            .map(|extensions| extensions.extension_canister_ids)
-            .unwrap_or_default();
-
-        Ok(extensions)
-    }
-
-    async fn canister_module_hash(
-        &self,
-        canister_id: CanisterId,
-    ) -> Result<Vec<u8>, GovernanceError> {
-        let canister_info_arg =
-            Encode!(&CanisterInfoRequest::new(canister_id, Some(1),)).map_err(|err| {
-                GovernanceError::new_with_message(
-                    ErrorType::External,
-                    format!("Error encoding canister_info request.\n{}", err),
-                )
-            })?;
-
-        let response = self
-            .env
-            .call_canister(CanisterId::ic_00(), "canister_info", canister_info_arg)
-            .await
-            .map_err(|err: (Option<i32>, String)| {
-                GovernanceError::new_with_message(
-                    ErrorType::External,
-                    format!("Canister method call IC00.canister_info failed: {:?}", err),
-                )
-            })
-            .and_then(|blob| {
-                Decode!(&blob, CanisterInfoResponse).map_err(|err| {
-                    GovernanceError::new_with_message(
-                        ErrorType::External,
-                        format!("Error decoding IC00.canister_info response:\n{}", err),
-                    )
-                })
-            })?;
-
-        Ok(response.module_hash().unwrap_or_default())
-    }
-
     async fn perform_execute_extension_operation(
         &self,
         execute_extension_operation: ExecuteExtensionOperation,
@@ -2694,63 +2621,16 @@ impl Governance {
             )
         })?;
 
-        let registered_extensions = self.list_extensions().await?;
-
-        if !registered_extensions.contains(&extension_canister_id.get()) {
-            return Err(GovernanceError::new_with_message(
-                ErrorType::NotFound,
-                format!(
-                    "Extension canister {} is not registered with the SNS.",
-                    extension_canister_id
-                ),
-            ));
-        }
-
-        let wasm_module_hash = self.canister_module_hash(extension_canister_id).await?;
-
-        let (extension_kind, extension_operations) =
-            match validate_extension_wasm(&wasm_module_hash) {
-                Ok(ExtensionSpec {
-                    kind, operations, ..
-                }) => (kind, operations),
-                Err(err) => {
-                    return Err(GovernanceError::new_with_message(
-                        ErrorType::InvalidProposal,
-                        format!(
-                            "Extension canister {} does not have an extension spec despite being \
-                         registered with Root: {}",
-                            extension_canister_id, err,
-                        ),
-                    ));
-                }
-            };
-
-        if extension_kind != ExtensionKind::TreasuryManager {
-            return Err(GovernanceError::new_with_message(
-                ErrorType::InvalidProposal,
-                "Only TreasuryManager extensions are currently supported.",
-            ));
-        }
+        validate_execute_extension_operation(
+            &*self.env,
+            self.proto.root_canister_id_or_panic(),
+            extension_canister_id,
+            operation_name,
+            &operation_arg,
+        )
+        .await?;
 
         let treasury_manager_canister_id = extension_canister_id;
-
-        let Some(ExtensionOperationSpec { name, .. }) = extension_operations.get(&operation_name)
-        else {
-            return Err(GovernanceError::new_with_message(
-                ErrorType::InvalidProposal,
-                format!(
-                    "Extension canister {} does not have an operation named {}",
-                    extension_canister_id, operation_name
-                ),
-            ));
-        };
-
-        if name != "deposit" {
-            return Err(GovernanceError::new_with_message(
-                ErrorType::InvalidProposal,
-                "Only TreasuryManager.deposit extensions are currently supported.",
-            ));
-        }
 
         let (arg, sns_amount_e8s, icp_amount_e8s) =
             self.construct_treasury_manager_deposit_payload(operation_arg)?;
