+++ rs/sns/governance/src/proposal.rs
@@ -1,5 +1,6 @@
 use crate::cached_upgrade_steps::render_two_versions_as_markdown_table;
-use crate::extensions::validate_extension_wasm;
+use crate::extensions::validate_execute_extension_operation;
+use crate::extensions::{validate_extension_wasm, ValidatedExecuteExtensionOperation};
 use crate::pb::v1::{
     AdvanceSnsTargetVersion, ExecuteExtensionOperation, RegisterExtension,
     SetTopicsForCustomProposals, Topic,
@@ -447,8 +448,8 @@ pub(crate) async fn validate_and_render_action(
             validate_and_render_execute_nervous_system_function(env, execute, existing_functions)
                 .await
         }
-        proposal::Action::ExecuteExtensionOperation(execute_extension_operation) => {
-            validate_and_render_execute_extension_operation(env, execute_extension_operation).await
+        proposal::Action::ExecuteExtensionOperation(execute) => {
+            validate_and_render_execute_extension_operation(env, execute, root_canister_id).await
         }
         proposal::Action::RegisterDappCanisters(register_dapp_canisters) => {
             validate_and_render_register_dapp_canisters(
@@ -1506,10 +1507,36 @@ pub async fn validate_and_render_execute_nervous_system_function(
 }
 
 async fn validate_and_render_execute_extension_operation(
-    _env: &dyn Environment,
-    _execute_extension_operation: &ExecuteExtensionOperation,
+    env: &dyn Environment,
+    execute: &ExecuteExtensionOperation,
+    root_canister_id: CanisterId,
 ) -> Result<String, String> {
-    Err("ExecuteExtensionOperation is not supported yet.".to_string())
+    let ValidatedExecuteExtensionOperation {
+        extension_canister_id,
+        operation_name,
+        operation_arg,
+    } = execute.clone().try_into()?;
+
+    validate_execute_extension_operation(
+        env,
+        root_canister_id,
+        extension_canister_id,
+        operation_name.clone(),
+        &operation_arg,
+    )
+    .await
+    .map_err(|err| err.error_message)?;
+
+    let operation_arg = format!("{:#?}", operation_arg);
+
+    Ok(format!(
+        r"# Proposal to execute extension operation:
+
+* Extension canister ID: `{extension_canister_id}`
+* Operation name: `{operation_name}`
+* Operation argument: `{operation_arg}`
+#"
+    ))
 }
 
 async fn validate_and_render_register_extension(
